[ADD] Ability to save processed datasets
